修改Identity.h文件名称，在system.cpp中优化for循环中的i为size_t类型，并创建了Makefile文件

+++ system.cpp
@@ -148,7 +148,7 @@ void System::Calculate_Rank(int select)
 		classOne_rank = 0;
 	}
 
-	for (int i = 1; i < v_Stu.size(); ++i)
+	for (size_t i = 1; i < v_Stu.size(); ++i)
 	{
 		if (v_Stu[i].getSumScore() == v_Stu[i - 1].getSumScore())
 		{
